Introduction to Machine Learning(CENG 499)/hw1/499Hw1ArtificialNeuralNetwork/requirements.txt
arcade==2.3.14
attrs==19.3.0
cloudpickle==1.3.0
dataclasses==0.7
future==0.18.2
gridworld==0.1
gym==0.17.1
install==1.3.4
numpy==1.18.2
Pillow==7.1.1
pip==20.2.4
pyglet==1.5.0
pytiled-parser==0.9.3
scipy==1.4.1
setuptools==39.0.1
six==1.14.0
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
virtualenv==16.7.8
